fix crash bug when multiple listeners disconnect at once
Vec::swap_remove removes an element from a Vec in constant time by replacing
it with the last element of the Vec. This can cause an out-of-bounds removal
if a to-be-removed element was previously swapped to another index. Note that
the buggy code never removed the _wrong_ element, since in every swap_remove
operation only the last element in the vec would be swapped to an index that
we won't ever visit again.

Removing elements starting from the greatest index first guarantees that
we'll never end up with an index greater than the length of the vec.